// ASEDINFO_FRONTEND/src/app/main/pages/compartidos/interfaces/IusuarioCedula.ts
export interface IusuarioCedula {
    cedula: string;
    cedulaMadre: string;
    cedulaPadre: string;
    codCondicionCedulado: number;
    codDomicilio: number;
    codEstadoCivil: number;
    codInstruccion: number;
    codLugarMatrimonio: number;
    codLugarNacimiento: number;
    codNacionalidad: number;
    codProfesion: string;
    codSexo: number;
    fechaExpedicionCed: string;
    fechaFallecimiento: string;
    fechaMatrimonio: string;
    fechaNacimiento: string;
    nombreCalle: string;
    nombreConyuge: string;
    nombreMadre: string;
    nombrePadre: string;
    nombres: string;
    numeroCasa: string;
}